+++ lib/chef/provider/link.rb
@@ -22,6 +22,7 @@ require 'chef/mixin/file_class'
 require 'chef/resource/link'
 require 'chef/provider'
 require 'chef/scan_access_control'
+require 'chef/util/path_helper'
 
 class Chef
   class Provider