+++ lib/chef/provider/service/macosx.rb
@@ -19,6 +19,7 @@
 require 'rexml/document'
 require 'chef/resource/service'
 require 'chef/provider/service/simple'
+require 'chef/util/path_helper'
 
 class Chef
   class Provider
@@ -194,7 +195,7 @@ class Chef
           plists = PLIST_DIRS.inject([]) do |results, dir|
             edir = ::File.expand_path(dir)
             entries = Dir.glob(
-              "#{edir}/*#{@current_resource.service_name}*.plist"
+              "#{edir}/*#{Chef::Util::PathHelper.escape_glob(@current_resource.service_name)}*.plist"
             )
             entries.any? ? results << entries : results
           end