+++ CHANGELOG.md
@@ -75,6 +75,9 @@
 
 ### Chef Contributions
 
+* Removed shelling out to erubis/ruby for syntax checks (>= 1.9 has been able
+  to do this in the ruby vm itself for awhile now and we've dropped 1.8.7 which
+  could not do this and had to shell_out)
 * Report the request and response when a non-200 error code happens
 * [FEATURE] Upgrade `knife upload` and `knife download` to download
   **everything** in an organization, now including the organization definition
@@ -135,6 +138,7 @@
 * Transfer trusted certificates under :trusted_certs_dir during bootstrap.
 * Set :ssl_verify_mode to :verify_peer by default.
 * Add homebrew provider for package resource, use it by default on OS X (Issue #1709)
+* Add escape_glob method to PathHelper, update glob operations.
 * Verify x509 properties of certificates in the :trusted_certs_dir during knife ssl check.
 
 ## Last Release: 11.14.2